1.1.0 - add configuration object
