+++ lib/ryquest/version.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# Root module
 module Ryquest
   # Ryquest version number: major.minor.maintenance
   VERSION = '1.0.0'
